Start doing renderer.
Renderer is a library ought to be able to render poligons as
ascii with specific flavours. It should also make the rendered stuff
scalable to a specific sized canvas.
For example you create a rectangle, and say that it's flavour should be
cloud, then the renderer will try to output somthing cloudy based on the
rectangle you gave.
 _._
(  °) <-- don't know yet how to create something cloudy (;_;)
 `'´      but it sort of looks like a jellyfish hm... interresting.